wip: refactor into multiple files
