+++ Sources/Application/ESP32AppROM.swift
@@ -1,0 +1,2 @@
+@_silgen_name("esp_app_desc")
+func _getAppDesc() -> UnsafeRawPointer
