+++ Sources/Application/ESP32DelayROM.swift
@@ -1,0 +1,2 @@
+@_silgen_name("esp_rom_delay_us")
+func esp_rom_delay_us(_ us: UInt32)
