+++ Sources/Application/ESP32UARTROM.swift
@@ -1,0 +1,5 @@
+@_silgen_name("esp_rom_uart_putc")
+func esp_rom_uart_putc(_ char: UInt8)
+
+@_silgen_name("esp_rom_uart_tx_wait_idle")
+func esp_rom_uart_tx_wait_idle(_ uart_no: UInt8)
